sam/functions/create_index/requirements.txt
elasticsearch-dsl==7.4.1
aws-lambda-powertools[tracer]==2.15.0